+++ changelog.md
@@ -1,8 +1,15 @@
+# v0.3.0
+
+- Change desktop extension UI
+- Add Home button with a small example of how to deploy workload
+
 # v0.2.0
+
 - Open port to the mesos agent
 - Add simple webui functionality
 - Update mini Apache Mesos to version 1.11.0-0.2.0
 - Add show details of frameworks, tasks and agents
 
 # v0.1.0
+
 - First inn